refactor: 図エディタ側のDragAndDropValueを移動。
