+++ src/components/diagrams/editor/template/DragAndDropValue.ts
